requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.1.0
numpy>=1.24.0
openpyxl>=3.1.0
dateparser>=1.2.0
regex>=2023.10.3

# Optional: OCR support for scanned PDFs
# Install these for OCR functionality (also requires Tesseract OCR to be installed on your system)
# pytesseract>=0.3.10
# pdf2image>=1.16.3
# Pillow>=10.0.0

